backend/requirements.txt
asgiref==3.2.7
atomicwrites==1.4.0
attrs==19.3.0
Django==2.2.12
djangorestframework==3.8.2
importlib-metadata==1.6.0
more-itertools==8.2.0
packaging==20.3
pluggy==0.6.0
py==1.8.1
pyparsing==2.4.7
psycopg2==2.7.4
pytest==3.6.1
pytest-django==3.2.1
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
wcwidth==0.1.9
zipp==3.1.0
